calendar_index/pyproject.toml
[tool.poetry]
name = "calendar-index"
version = "0.1.0"
description = ""
authors = ["Damian Akpan <damiakpan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
prefect = "^2.16.9"
trafilatura = "^1.8.1"
openai = "^1.16.2"
langchain-text-splitters = "^0.0.1"
anthropic = "^0.23.1"
tenacity = "^8.2.3"
duckdb = "^0.10.1"
wikipedia = "^1.4.0"
google-generativeai = "^0.4.1"
pydantic = "^2.6.4"
psycopg2-binary = "^2.9.9"
httpx = {extras = ["socks"], version = "^0.27.0"}
selenium = "^4.19.0"
numpy = "^1.26.4"
tavily-python = "^0.3.3"
nltk = "^3.8.1"
warc = "^0.2.1"
pandas = "^2.2.2"
nanoid = "^2.0.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.7"
pylint = "^3.1.0"
prefect = "^2.17.1"
ipython = "^8.23.0"
b2 = "^3.19.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
